Refactor #349 - make annotation module for ExcludeFromGeneratedTestReport
