+++ app/build.gradle.kts
@@ -77,6 +77,7 @@ dependencies {
     implementation(projects.core.androidx.lifecycle)
     implementation(projects.core.androidx.recyclerview)
     implementation(projects.core.androidx.transition)
+    implementation(projects.core.annotation)
     implementation(projects.core.component.tag)
     implementation(projects.core.component.uitext)
     implementation(projects.core.dataDi)
@@ -84,7 +85,6 @@ dependencies {
     implementation(projects.core.designsystem)
     implementation(projects.core.domain)
     implementation(projects.core.domainDi)
-    implementation(projects.core.foundation)
     implementation(projects.core.inject)
     implementation(projects.core.kotlin)
     implementation(projects.core.kotlinxCoroutine)
