+++ core/annotation/build.gradle.kts
@@ -1,4 +1,3 @@
 plugins {
     alias(libs.plugins.nlab.jvm.library)
-    alias(libs.plugins.nlab.jvm.library.jacoco)
 }
