+++ core/component/tag/build.gradle.kts
@@ -10,6 +10,7 @@ android {
 }
 
 dependencies {
+    implementation(projects.core.annotation)
     implementation(projects.core.dataDi)
     implementation(projects.core.domainDi)
     implementation(projects.core.translation)
