+++ core/data-impl/build.gradle.kts
@@ -8,8 +8,8 @@ android {
 }
 
 dependencies {
+    implementation(projects.core.annotation)
     implementation(projects.core.data)
-    implementation(projects.core.foundation)
     implementation(projects.core.kotlinxCoroutine)
     implementation(projects.core.local)
     implementation(projects.core.network)
