+++ core/data-impl/src/main/kotlin/com/nlab/reminder/core/data/repository/fake/FakeScheduleRepositoryDelegate.kt
@@ -16,7 +16,7 @@
 
 package com.nlab.reminder.core.data.repository.fake
 
-import com.nlab.reminder.core.foundation.annotation.ExcludeFromGeneratedTestReport
+import com.nlab.reminder.core.annotation.ExcludeFromGeneratedTestReport
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
