+++ settings.gradle.kts
@@ -56,6 +56,7 @@ include(
     ":core:androidx:lifecycle",
     ":core:androidx:recyclerview",
     ":core:androidx:transition",
+    ":core:annotation",
     ":core:component:tag",
     ":core:component:tag-test",
     ":core:component:uitext",
@@ -68,7 +69,6 @@ include(
     ":core:designsystem",
     ":core:domain",
     ":core:domain-di",
-    ":core:foundation",
     ":core:inject",
     ":core:kotlin",
     ":core:kotlin-test",
